LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/windows/browsing-context-names/choose-_blank-003-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT Context created by link targeting "_blank" should not have opener reference Test timed out
Link
